refactor: split user keyboard builders
